+++ src/widget/box_widget.rs
@@ -25,17 +25,17 @@ use crate::widget::widget::*;
 pub struct BoxWidget {
     config: Configurable,
     callbacks: CallbackStore,
-    base_widget: BaseWidget,
+    base_widget: CanvasWidget,
 }
 
-/// Implementation of the constructor for the `BaseWidget`.  Creates a new base widget
+/// Implementation of the constructor for the `CanvasWidget`.  Creates a new base widget
 /// that can be positioned anywhere on the screen.
 impl BoxWidget {
     pub fn new() -> Self {
         Self {
             config: Configurable::new(),
             callbacks: CallbackStore::new(),
-            base_widget: BaseWidget::new(),
+            base_widget: CanvasWidget::new(),
         }
     }
 
@@ -135,8 +135,8 @@ impl BoxWidget {
 }
 
 /// Implementation of the `BoxWidget` object with the `Widget` traits implemented.
-/// This implementation is similar to the `BaseWidget`, but incorporates a drawable box inside
-/// the widget.  Base widget is the `BaseWidget`.
+/// This implementation is similar to the `CanvasWidget`, but incorporates a drawable box inside
+/// the widget.  Base widget is the `CanvasWidget`.
 ///
 /// This is basically just a box with a fill color.  Use this to draw other things like buttons,
 /// text widgets, and so on, if you need anything with a drawable border.